refractor Header component to own sub-folder with children components
